+++ pkg/backend/livestream/youtube/notif.go
@@ -18,7 +18,7 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-func (PushData *NotifStruct) SendNude() {
+func (PushData *NotifStruct) SendNude() error {
 	Status := PushData.YtData.Status
 	Avatar := PushData.Member.YoutubeAvatar
 	YtChannel := "https://www.youtube.com/channel/" + PushData.Member.YoutubeID + "?sub_confirmation=1"
@@ -55,7 +55,7 @@ func (PushData *NotifStruct) SendNude() {
 	if Status == "upcoming" {
 		Color, err := engine.GetColor(config.TmpDir, PushData.YtData.Thumb)
 		if err != nil {
-			log.Error(err)
+			return err
 		}
 		//id, DiscordChannelID
 		var (
@@ -64,13 +64,13 @@ func (PushData *NotifStruct) SendNude() {
 		)
 		for i, v := range ChannelData {
 			wg.Add(1)
-			go func(Channel database.DiscordChannel, wg *sync.WaitGroup) {
+			go func(Channel database.DiscordChannel, wg *sync.WaitGroup) error {
 				defer wg.Done()
 				UserTagsList := Channel.GetUserList(context.Background()) //database.GetUserList(Channel.ID, PushData.Member.ID)
 				if UserTagsList == nil && PushData.Group.GroupName != "Independen" {
 					UserTagsList = []string{"_"}
 				} else if UserTagsList == nil && PushData.Group.GroupName == "Independen" {
-					return
+					return nil
 				}
 				msg, err := Bot.ChannelMessageSendEmbed(Channel.ChannelID, engine.NewEmbed().
 					SetAuthor(VtuberName, Avatar, YtChannel).
@@ -86,16 +86,21 @@ func (PushData *NotifStruct) SendNude() {
 					SetFooter(Timestart.In(loc).Format(time.RFC822), config.YoutubeIMG).
 					SetColor(Color).MessageEmbed)
 				if err != nil {
-					log.Error(msg, err)
+					log.WithFields(log.Fields{
+						"Message":          msg,
+						"ChannelID":        Channel.ID,
+						"DiscordChannelID": Channel.ChannelID,
+					}).Error(err)
 					err = Channel.DelChannel(err.Error())
 					if err != nil {
-						log.Error(err)
+						return err
 					}
 				}
 				msg, err = Bot.ChannelMessageSend(Channel.ChannelID, "`"+PushData.Member.Name+"` New upcoming Livestream\nUserTags: "+strings.Join(UserTagsList, " "))
 				if err != nil {
-					log.Error(err)
+					return err
 				}
+				return nil
 			}(v, &wg)
 			//Wait every ge 10 discord channel
 			if i%10 == 0 {
@@ -107,13 +112,13 @@ func (PushData *NotifStruct) SendNude() {
 	} else if Status == "live" {
 		Color, err := engine.GetColor(config.TmpDir, PushData.YtData.Thumb)
 		if err != nil {
-			log.Error(err)
+			return err
 		}
 		Bili := false
 		if PushData.Member.BiliRoomID != 0 {
 			LiveBili, err := live.GetRoomStatus(PushData.Member.BiliRoomID)
 			if err != nil {
-				log.Error(err)
+				return err
 			}
 			if LiveBili.CheckScheduleLive() {
 				Bili = true
@@ -127,13 +132,13 @@ func (PushData *NotifStruct) SendNude() {
 		)
 		for i, v := range ChannelData {
 			wg.Add(1)
-			go func(Channel database.DiscordChannel, wg *sync.WaitGroup) {
+			go func(Channel database.DiscordChannel, wg *sync.WaitGroup) error {
 				defer wg.Done()
 				UserTagsList := Channel.GetUserList(context.Background()) //database.GetUserList(Channel.ID, PushData.Member.ID)
 				if UserTagsList == nil && PushData.Group.GroupName != "Independen" {
 					UserTagsList = []string{"_"}
 				} else if UserTagsList == nil && PushData.Group.GroupName == "Independen" {
-					return
+					return nil
 				}
 				MsgEmbed, err := Bot.ChannelMessageSendEmbed(Channel.ChannelID, engine.NewEmbed().
 					SetAuthor(VtuberName, Avatar, YtChannel).
@@ -149,60 +154,64 @@ func (PushData *NotifStruct) SendNude() {
 					SetFooter(Timestart.In(loc).Format(time.RFC822), config.YoutubeIMG).
 					SetColor(Color).MessageEmbed)
 				if err != nil {
-					log.Error(MsgEmbed, err)
+					log.WithFields(log.Fields{
+						"Message":          MsgEmbed,
+						"ChannelID":        Channel.ID,
+						"DiscordChannelID": Channel.ChannelID,
+					}).Error(err)
 					err = Channel.DelChannel(err.Error())
 					if err != nil {
-						log.Error(err)
-					}
-				} else {
-					if Channel.Dynamic {
-						log.WithFields(log.Fields{
-							"DiscordChannel": Channel.ChannelID,
-							"VtuberGroupID":  PushData.Group.ID,
-							"YoutubeID":      PushData.YtData.ID,
-						}).Info("Set dynamic mode")
-						Channel.SetVideoID(PushData.YtData.VideoID).
-							SetMsgEmbedID(MsgEmbed.ID)
-					}
-					Msg := "Push " + config.BotConf.Emoji.Livestream[0] + " to add you in `" + PushData.Member.Name + "` ping list\nPush " + config.BotConf.Emoji.Livestream[1] + " to remove you from ping list"
-					MsgID := ""
-					if Bili {
-						msg, err := Bot.ChannelMessageSend(Channel.ChannelID, "`"+PushData.Member.Name+"` Live right now at BiliBili And Youtube\nUserTags: "+strings.Join(UserTagsList, " ")+"\n"+Msg)
-						if err != nil {
-							log.Error(err)
-						}
-						if Channel.Dynamic {
-							Channel.SetMsgTextID(msg.ID).PushReddis()
-						}
-						MsgID = msg.ID
-
-					} else {
-						msg, err := Bot.ChannelMessageSend(Channel.ChannelID, "`"+PushData.Member.Name+"` Live right now\nUserTags: "+strings.Join(UserTagsList, " ")+"\n"+Msg)
-						if err != nil {
-							log.Error(err)
-						}
-						if Channel.Dynamic {
-							Channel.SetMsgTextID(msg.ID).PushReddis()
-						}
-						MsgID = msg.ID
-					}
-					User.SetDiscordChannelID(Channel.ChannelID).
-						SetGroup(PushData.Group).
-						SetMember(PushData.Member)
-					err = User.SendToCache(MsgID)
-					if err != nil {
-						log.Error(err)
-					}
-
-					err = engine.Reacting(map[string]string{
-						"ChannelID": Channel.ChannelID,
-						"State":     "Youtube",
-						"MessageID": MsgID,
-					}, Bot)
-					if err != nil {
-						log.Error(err)
+						return err
 					}
 				}
+				if Channel.Dynamic {
+					log.WithFields(log.Fields{
+						"DiscordChannel": Channel.ChannelID,
+						"VtuberGroupID":  PushData.Group.ID,
+						"YoutubeID":      PushData.YtData.ID,
+					}).Info("Set dynamic mode")
+					Channel.SetVideoID(PushData.YtData.VideoID).
+						SetMsgEmbedID(MsgEmbed.ID)
+				}
+				Msg := "Push " + config.BotConf.Emoji.Livestream[0] + " to add you in `" + PushData.Member.Name + "` ping list\nPush " + config.BotConf.Emoji.Livestream[1] + " to remove you from ping list"
+				MsgID := ""
+				if Bili {
+					msg, err := Bot.ChannelMessageSend(Channel.ChannelID, "`"+PushData.Member.Name+"` Live right now at BiliBili And Youtube\nUserTags: "+strings.Join(UserTagsList, " ")+"\n"+Msg)
+					if err != nil {
+						return err
+					}
+					if Channel.Dynamic {
+						Channel.SetMsgTextID(msg.ID).PushReddis()
+					}
+					MsgID = msg.ID
+
+				} else {
+					msg, err := Bot.ChannelMessageSend(Channel.ChannelID, "`"+PushData.Member.Name+"` Live right now\nUserTags: "+strings.Join(UserTagsList, " ")+"\n"+Msg)
+					if err != nil {
+						return err
+					}
+					if Channel.Dynamic {
+						Channel.SetMsgTextID(msg.ID).PushReddis()
+					}
+					MsgID = msg.ID
+				}
+				User.SetDiscordChannelID(Channel.ChannelID).
+					SetGroup(PushData.Group).
+					SetMember(PushData.Member)
+				err = User.SendToCache(MsgID)
+				if err != nil {
+					return err
+				}
+
+				err = engine.Reacting(map[string]string{
+					"ChannelID": Channel.ChannelID,
+					"State":     "Youtube",
+					"MessageID": MsgID,
+				}, Bot)
+				if err != nil {
+					return err
+				}
+				return nil
 			}(v, &wg)
 			//Wait every ge 15 discord channel
 			if i%15 == 0 {
@@ -214,7 +223,7 @@ func (PushData *NotifStruct) SendNude() {
 	} else if Status == "past" {
 		Color, err := engine.GetColor(config.TmpDir, PushData.YtData.Thumb)
 		if err != nil {
-			log.Error(err)
+			return err
 		}
 
 		//id, DiscordChannelID
@@ -224,7 +233,7 @@ func (PushData *NotifStruct) SendNude() {
 		)
 		for i, v := range ChannelData {
 			wg.Add(1)
-			go func(Channel database.DiscordChannel, wg *sync.WaitGroup) {
+			go func(Channel database.DiscordChannel, wg *sync.WaitGroup) error {
 				defer wg.Done()
 				UserTagsList := Channel.GetUserList(context.Background()) //database.GetUserList(Channel.ID, PushData.Member.ID)
 				if UserTagsList != nil {
@@ -243,17 +252,22 @@ func (PushData *NotifStruct) SendNude() {
 						SetFooter(Timestart.In(loc).Format(time.RFC822), config.YoutubeIMG).
 						SetColor(Color).MessageEmbed)
 					if err != nil {
-						log.Error(msg, err)
+						log.WithFields(log.Fields{
+							"Message":          msg,
+							"ChannelID":        Channel.ID,
+							"DiscordChannelID": Channel.ChannelID,
+						}).Error(err)
 						err = Channel.DelChannel(err.Error())
 						if err != nil {
-							log.Error(err)
+							return err
 						}
 					}
 					msg, err = Bot.ChannelMessageSend(Channel.ChannelID, "`"+PushData.Member.Name+"` Uploaded a new video\nUserTags: "+strings.Join(UserTagsList, " "))
 					if err != nil {
-						log.Error(err)
+						return err
 					}
 				}
+				return nil
 			}(v, &wg)
 			//Wait every ge 10 discord channel
 			if i%10 == 0 {
@@ -266,9 +280,8 @@ func (PushData *NotifStruct) SendNude() {
 		var Color int
 		ChanelData := database.ChannelTag(PushData.Member.ID, 2, "")
 		for _, Channel := range ChanelData {
-			for ii := 10; ii < 70; ii += 5 {
-				k := ii - 6
-				if UpcominginMinutes <= ii && UpcominginMinutes >= k {
+			for ii := 0; ii < 70; ii++ {
+				if UpcominginMinutes == ii {
 					if Color != 0 {
 						Color, err = engine.GetColor(config.TmpDir, PushData.YtData.Thumb)
 						if err != nil {
@@ -293,15 +306,17 @@ func (PushData *NotifStruct) SendNude() {
 							SetFooter(Timestart.In(loc).Format(time.RFC822), config.YoutubeIMG).
 							SetColor(Color).MessageEmbed)
 						if err != nil {
-							log.Error(MsgEmbed, err)
+							log.WithFields(log.Fields{
+								"Message":          MsgEmbed,
+								"ChannelID":        Channel.ID,
+								"DiscordChannelID": Channel.ChannelID,
+							}).Error(err)
 							err = Channel.DelChannel(err.Error())
-							if err != nil {
-								log.Error(err)
-							}
+							return err
 						}
 						MsgText, err := Bot.ChannelMessageSend(Channel.ChannelID, "`"+PushData.Member.Name+"` Live in "+LiveCount+"\nUserTags: "+strings.Join(UserTagsList, " "))
 						if err != nil {
-							log.Error(err)
+							return err
 						}
 						if Channel.Dynamic {
 							log.WithFields(log.Fields{
@@ -318,4 +333,5 @@ func (PushData *NotifStruct) SendNude() {
 			}
 		}
 	}
+	return nil
 }