+++ pkg/frontend/bilibili_handler.go
@@ -1,4 +1,4 @@
-package discordhandler
+package main
 
 import (
 	"strconv"