+++ pkg/frontend/subscriber_handler.go
@@ -1,4 +1,4 @@
-package discordhandler
+package main
 
 import (
 	"math/rand"