+++ tools/config/config.go
@@ -53,7 +53,7 @@ const (
 	BiliBiliLive               = "@every 0h7m0s"
 	BiliBiliSpace              = "@every 0h13m0s"
 	YoutubeCheckChannel        = "@every 0h10m30s"
-	YoutubeCheckUpcomingByTime = "@every 0h2m0s"
+	YoutubeCheckUpcomingByTime = "@every 0h1m0s"
 	YoutubePrivateSlayer       = "@every 2h31m0s"
 	YoutubeSubscriber          = "@every 1h0m0s"
 	BiliBiliFollowers          = "@every 0h30m0s"